+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.1.8
+
+* OpenTelemetry tracing via `WithOpenTelemetryTracing` http.ServerOption.
+* TLS support via `WithTLS` http.ServerOption.
+
 ## v0.1.7
 
 * Metrics improvements